move functions_registry into daft-dsl
